+++ mapmatching/osmnet/parse_osm_xml.py
@@ -10,12 +10,10 @@ from shapely.geometry import Point, LineString
 
 warnings.filterwarnings('ignore')
 
-import sys
 from ..utils.timer import Timer, timeit
 from ..osmnet.misc import Bunch, cal_od_straight_distance
-from ..osmnet.osm_io import load_graph, save_graph
 from ..osmnet.twoway_edge import add_reverse_edge, swap_od, edge_offset
-from ..osmnet.combine_edges import pipeline_combine_links
+from ..osmnet.combine_edges import pipeline_combine_links, parallel_process
 from ..setting import highway_filters, link_type_level_dict
 from ..geo.haversineDistance import points_geoseries_2_ndarray
 
@@ -39,10 +37,11 @@ class WayHandler(SimpleHandler):
         if self.link_type_filter is not None and way.highway in self.link_type_filter:
             return
 
-        # ways
+        # ways # TODO filter key params
         way['osm_way_id'] = int(w.id)
         way['ref_node_id_list'] = [int(node.ref) for node in w.nodes]
         way['name'] = w.tags.get('name')
+        way['service'] = w.tags.get('service')
         self._extract_lane_info(w, way)
         self._extract_maxspeed_info(w, way)
         self._extract_one_way_info(w, way)
@@ -52,15 +51,14 @@ class WayHandler(SimpleHandler):
 
         # edges
         nds = way.ref_node_id_list
-        for i in range(len(way.ref_node_id_list)-1):
+        for i in range(len(way.ref_node_id_list) - 1):
             item = {'way_id': way['osm_way_id'], 
                     'order': i, 
-                    'src':nds[i], 
-                    'dst':nds[i+1],
-                    'waypoints': f"[{nds[i]},{nds[i+1]}]"
+                    'src': nds[i], 
+                    'dst': nds[i+1],
+                    'waypoints': f"[{nds[i]},{nds[i+1]}]" 
             }
             self.edges.append(item)
-
 
     def _extract_one_way_info(self, w, way):
         # TODO check and update one way info by the `detour coefficient`
@@ -83,7 +81,6 @@ class WayHandler(SimpleHandler):
             way['oneway'] = True
         
         return way
-
 
     def _extract_lane_info(self, w, way):
         lane_info = w.tags.get('lanes')
@@ -108,7 +105,6 @@ class WayHandler(SimpleHandler):
 
         return way
 
-    
     def _extract_maxspeed_info(self, w, way):
         maxspeed_info = w.tags.get('maxspeed')
         way['maxspeed'] = None
@@ -165,23 +161,28 @@ def check_multi_edges(df_edges):
         return []
         
     idxs = set(np.concatenate(idxs))
-
-    status = len(idxs) > 0
-    if status:
-        print(f"Exists {len(idxs)} multi-egdes: \n\t{idxs}")
+    if len(idxs):
+        logger.debug(f"Exists {len(idxs)} multi-egdes: \n\t{idxs}")
         return list(idxs)
     
     return []
 
-def _extract_od(points):
+def __extract_waypoints_2_od(points):
     if isinstance(points, str):
         points = eval(points)
     assert isinstance(points, list), "check od type"
     
     return [str([points[i], points[i+1]]) for i in range(len(points)-1)]
 
-def post_process_ways(way_handler):
-    df_ways = pd.DataFrame.from_dict(way_handler.osm_way_dict, orient='index')
+def __check_dulplicates(df):
+    # BUG
+    src, dst = 10253667641, 10253667620
+    tmp = df.query('src == @src and dst == @dst')
+    if tmp.shape[0] > 1:
+        print("Exist duplicates !!!!")
+
+def post_process_ways(osm_way_dict):
+    df_ways = pd.DataFrame.from_dict(osm_way_dict, orient='index')
     df_ways.loc[:, 'link'] = df_ways.highway.apply(lambda x: "link" in x)
     df_ways.loc[:, 'highway'] = df_ways.highway.apply(lambda x: x.split('_')[0])
     df_ways.loc[:, 'level'] = df_ways.highway.apply(lambda x: link_type_level_dict.get(x, 99))
@@ -199,26 +200,25 @@ def drop_duplicates_way(df, df_ways=None, att="ref_node_id_list", sort_atts=['sr
     idxs = lst.to_frame().query(f"{att} in @contents").index
     _df = df.loc[idxs]
     if df_ways:
-        _df = append_level_to_df(_df, df_ways)
+        _df = append_way_info(_df, df_ways, ['level'])
 
     duplicates = _df.sort_values([i for i in sort_atts if i in list(df)])
     keep_idxs = set(duplicates.groupby(['src', 'dst']).head(1).index)
     drop_idxs = set([i for i in duplicates.index if i not in keep_idxs])
 
-    # drop_ways = df_ways.loc[drop_idxs]
     df.drop(index=drop_idxs, inplace=True)
 
     return df, keep_idxs, drop_idxs
 
-def append_level_to_df(df, df_ways, left_on='way_id'):
-    return df.merge(df_ways[['level']], left_on=left_on, right_index=True, how='left')
-
-def check_dulplicates(df):
-    # BUG
-    src, dst = 10253667641, 10253667620
-    tmp = df.query('src == @src and dst == @dst')
-    if tmp.shape[0] > 1:
-        print("Exist duplicates !!!!")
+def append_way_info(df_edges, df_ways, way_attrs=None):
+    if way_attrs is None:
+        way_attrs = list(df_ways)
+    
+    df_edges = df_edges.merge(df_ways[way_attrs], left_on='way_id', right_index=True, how='left')
+    if "highway" in way_attrs:
+        df_edges.rename(columns={'highway': 'road_type'}, inplace=True)
+    
+    return df_edges
 
 @timeit
 def _parse_xml(fn, highway_filters, crs="epsg:4326", in_sys='wgs', out_sys='wgs'):
@@ -226,10 +226,10 @@ def _parse_xml(fn, highway_filters, crs="epsg:4326", in_sys='wgs', out_sys='wgs'
     timer = Timer()
     way_handler = WayHandler(highway_filters)
     way_handler.apply_file(fn)
-    df_ways = post_process_ways(way_handler)
+    df_ways = post_process_ways(way_handler.osm_way_dict)
     df_ways, keep_idxs, drop_idxs = drop_duplicates_way(df_ways)    
     print(f"Parse ways: {timer.stop():.2f} s")
-    print(f"\tDrop duplicates ways: {drop_idxs}")
+    logger.debug(f"\tDrop duplicates ways, way ids: {list(drop_idxs)}")
 
     # nodes 
     timer.start()
@@ -249,22 +249,35 @@ def _parse_xml(fn, highway_filters, crs="epsg:4326", in_sys='wgs', out_sys='wgs'
         df_edges.loc[:, 'waypoints'] = df_edges.loc[:, 'waypoints'].apply(eval)
     if df_edges.shape[0] != _size:
         print(f"\tDrop duplicates ways / edges: {_size} -> {df_edges.shape[0]}")
-    check_dulplicates(df_edges)
 
     df_edges.loc[:, 'dist'] = cal_od_straight_distance(df_edges, df_nodes)
-    print(f"\tTransform to Dataframe: {timer.stop():.2f} s, node len: {df_nodes.shape[0]}")
+    print(f"Transform to Dataframe: {timer.stop():.2f} s, node len: {df_nodes.shape[0]}")
 
     return df_ways, df_nodes, df_edges
 
-@timeit
-def _transform_coords_seq_2_linstring(df_edges, df_nodes):
-    # Transform coords seq to geometry
-    df_edges.loc[:, 'waypoints'] = df_edges.waypoints.apply(
-            lambda x: eval(x) if isinstance(x, str) else x)
-    df_edges.loc[:,'geometry'] = df_edges.loc[:, 'waypoints'].apply(
+def _transform_coords_seq_2_linstring(waypoints, df_nodes):
+    # Execute in the as small as possible
+    coords = waypoints.apply(
+        lambda x: eval(x) if isinstance(x, str) else x)
+    geoms = coords.apply(
         lambda x: LineString(points_geoseries_2_ndarray(df_nodes.loc[x].geometry)))
     
-    return gpd.GeoDataFrame(df_edges)
+    return geoms
+
+@timeit
+def _parrallel_collect_geoms(df, df_nodes, n_jobs=16):
+    _size = df.shape[0] // n_jobs + 1
+    df.loc[:, 'part'] = df.index // _size
+
+    params = ((df.waypoints, df_nodes) for _, df in df.groupby('part'))
+    df.drop(columns=['part'], inplace=True)
+
+    geoms = parallel_process(_transform_coords_seq_2_linstring, params, pbar_switch=True, 
+                            n_jobs=n_jobs, total=n_jobs, desc='Collect coords')
+
+    geoms = pd.concat(geoms).sort_index()
+
+    return geoms
 
 @timeit
 def _simplify_edges(df_edges, df_nodes, n_jobs):
@@ -288,56 +301,55 @@ def _process_multi_edges(df_edges, ori_df_edges, df_ways):
     if "level" not in list(df_edges):
         df_edges = df_edges.merge(df_ways[['level']], left_on='way_id', right_index=True, how='left')
     
+    # (src, dst) 重复的记录
     idxs = check_multi_edges(df_edges)
+    normal_edges = df_edges.query("index not in @idxs")
     multi_edges = df_edges.loc[idxs].sort_values(
             ['src', 'dst', 'dist', 'level', 'way_id'], ascending=[True, True, True, True, True])
-    normal_edges = df_edges.query("index not in @idxs")
 
     # case 1: waypoints 一致的情况
-    multi_edges, keep_idxs_case1, drop_idxs_case1 = drop_duplicates_way(multi_edges, att='waypoints')
-    if drop_idxs_case1:
-        df_edges.loc[keep_idxs_case1.union(drop_idxs_case1)]
-        assert multi_edges.query(f"index == {list(drop_idxs_case1)[0]}").shape[0] == 0
-        print(f"Drop same waypoints: {drop_idxs_case1}")
+    multi_edges, keep_idxs_case1, case1_drop_idxs = drop_duplicates_way(multi_edges, att='waypoints')
+    if case1_drop_idxs:
+        df_edges.loc[keep_idxs_case1.union(case1_drop_idxs)]
+        assert multi_edges.query(f"index == {list(case1_drop_idxs)[0]}").shape[0] == 0
+        logger.debug(f"Drop same waypoints: {case1_drop_idxs}")
 
     # case 2: waypoints 内容不一致
     keep_idxs = multi_edges.groupby(['src', 'dst']).head(1).index
-    drop_idxs_case2 = set([i for i in idxs if i not in keep_idxs])
-    drop_idxs_case2.difference_update(drop_idxs_case1)
+    keeped_multi_edges = multi_edges.loc[keep_idxs]
+    case2_drop_idxs = set([i for i in idxs if i not in keep_idxs])
+    case2_drop_idxs.difference_update(case1_drop_idxs)
 
-    od_set = df_edges.loc[drop_idxs_case2].waypoints.apply(_extract_od).tolist()
-    od_set = set(np.concatenate(od_set))
-    df_edges_waypoints = ori_df_edges.waypoints.apply(str).to_frame()
+    if case2_drop_idxs:
+        od_set = df_edges.loc[case2_drop_idxs].waypoints.apply(__extract_waypoints_2_od).tolist()
+        od_set = set(np.concatenate(od_set))
+        df_edges_waypoints = ori_df_edges.waypoints.apply(str).to_frame()
 
-    pos = df_edges_waypoints.query("waypoints in @od_set")
-    pos_od_idxs = pos.index
-    pos_od_set = set(pos.waypoints.values)
-    pos_df_edges = ori_df_edges.loc[pos_od_idxs]
-    pos_df_edges.loc[:, 'dir'] = 1
+        pos = df_edges_waypoints.query("waypoints in @od_set")
+        pos_od_idxs = pos.index
+        pos_od_set = set(pos.waypoints.values)
+        pos_df_edges = ori_df_edges.loc[pos_od_idxs]
+        pos_df_edges.loc[:, 'dir'] = 1
 
-    neg_od_set = od_set - pos_od_set
-    neg_od_set = {str(eval(x)[::-1]) for x in neg_od_set}
-    neg_od_set = df_edges_waypoints.query("waypoints in @neg_od_set").index
-    neg_df_edges = swap_od(ori_df_edges.loc[neg_od_set])
-    neg_df_edges.loc[:, 'dir'] = -1
+        neg_od_set = od_set - pos_od_set
+        neg_od_set = {str(eval(x)[::-1]) for x in neg_od_set}
+        neg_od_set = df_edges_waypoints.query("waypoints in @neg_od_set").index
+        neg_df_edges = swap_od(ori_df_edges.loc[neg_od_set])
+        neg_df_edges.loc[:, 'dir'] = -1
 
-    _multi_edges = pd.concat([pos_df_edges, neg_df_edges])
-    print(f"od list len: {len(od_set)}, df_segs len: {_multi_edges.shape[0]}")
+        _multi_edges = pd.concat([pos_df_edges, neg_df_edges])
+        logger.debug(f"od list len: {len(od_set)}, df_segs len: {_multi_edges.shape[0]}")
+    else:
+        _multi_edges = None
 
     # case 3: 针对 waypoints 仅有两条记录的情况
-    ps = multi_edges.loc[drop_idxs_case2].waypoints
+    ps = multi_edges.loc[case2_drop_idxs].waypoints
     tmp = df_edges.loc[ps[ps.apply(len) == 2].index]
     assert tmp.shape[0] == 0
 
-    df_edges = pd.concat([normal_edges, multi_edges.loc[keep_idxs], _multi_edges])\
+    df_edges = pd.concat([normal_edges, keeped_multi_edges, _multi_edges])\
                  .drop(columns='level').reset_index(drop=True)
 
-    return df_edges
-
-def append_way_info(df_edges, df_ways, way_attrs = ['highway', 'level', 'name', 'oneway', 'link']):
-    df_edges = df_edges.merge(df_ways[way_attrs], left_on='way_id', right_index=True, how='left')
-    df_edges.rename(columns={'highway': 'road_type'}, inplace=True)
-    
     return df_edges
 
 def parse_xml_to_graph(fn, highway_filters=highway_filters, simplify=True, twoway=True, offset=True, 
@@ -348,13 +360,22 @@ def parse_xml_to_graph(fn, highway_filters=highway_filters, simplify=True, twowa
     ori_df_edges = df_edges_raw.copy()
     if simplify:
         df_edges = _simplify_edges(ori_df_edges, df_nodes, n_jobs=n_jobs)
+    
     if twoway:
         df_edges = _add_revert_edges(df_edges, df_ways)
+
     df_edges = _process_multi_edges(df_edges, ori_df_edges, df_ways)
-    df_edges = _transform_coords_seq_2_linstring(df_edges, df_nodes)
+
+    # df_edges = _transform_coords_seq_2_linstring(df_edges, df_nodes)
+    df_edges.loc[:, 'geometry'] = _parrallel_collect_geoms(df_edges[['waypoints']], df_nodes, n_jobs)
+    df_edges = gpd.GeoDataFrame(df_edges)
+
     if offset:
         df_edges = edge_offset(df_edges).sort_index()
-    df_edges = append_way_info(df_edges, df_ways)
+
+    way_attrs = ['name', 'level', 'highway', 'link', 'maxspeed', 'oneway', 'lanes', 'service']
+    df_edges = append_way_info(df_edges, df_ways, way_attrs)
+
     df_edges.reset_index(inplace=True, drop=True)
     df_edges.loc[:, 'eid'] = df_edges.index.values
 
@@ -386,5 +407,3 @@ if __name__ == "__main__":
     """ _add_reverse_edge """
     df_edges.query('way_id == 160131332').plot()
 
-
-# %%